Added more complex entity and updated forms, attached thymeleaf template to ease front-ending

+++ src/main/java/com/avgona/citizencrud/repository/CitizenRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CitizenRepository extends JpaRepository<Citizen, Integer> {
+public interface CitizenRepository extends JpaRepository<Citizen, Long> {
 
     // add a method to sort by last name
     public List<Citizen> findAllByOrderByLastNameAsc();
